review: test split into two files: kafka and req-resp
